requirements.txt
# Core ML Framework
torch>=2.0.0
torchaudio>=2.0.0
torchvision>=0.15.0

# Audio Processing
librosa>=0.9.0
soundfile>=0.12.0
scipy>=1.9.0

# Transformers
transformers>=4.20.0
tokenizers>=0.12.0

# Data Science
numpy>=1.21.0
pandas>=1.4.0
scikit-learn>=1.1.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Configuration
PyYAML>=6.0
omegaconf>=2.2.0

# Utilities
tqdm>=4.64.0
tensorboardX>=2.6
wandb>=0.12.0

# Development
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
isort>=5.10.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.0.0
ipywidgets>=7.7.0

# Audio Evaluation
pesq>=0.0.4
pystoi>=0.3.3

# Progress bars
rich>=12.0.0

# File handling
pathlib2>=2.3.0
